remove wrnings - sign compare
remove for "-Wsign-compare"

+++ include/quantileSamplingModule.hpp
@@ -50,7 +50,7 @@ public:
 		_completeTIs=completeTIs;
 
 		_useUniqueTI=useUniqueTI;
-		for (int i = 0; i < nbThread; ++i)
+		for (unsigned i = 0; i < nbThread; ++i)
 		{
 			_errors[i]=(float*)malloc(_cdmV[0].size() * int(ceil(_k)) * sizeof(float));
 			_encodedPosition[i]=(unsigned*)malloc(_cdmV[0].size() * int(ceil(_k)) * sizeof(unsigned));
@@ -63,13 +63,13 @@ public:
 		_convertionCoefVectorConstVector=convertionCoefVectorConstVector;
 	}
 	~QuantileSamplingModule(){
-		for (int i = 0; i < _errors.size(); ++i)
+		for (size_t i = 0; i < _errors.size(); ++i)
 		{
 			free(_errors[i]);
 			_errors[i]=nullptr;
 		}
 		
-		for (int i = 0; i < _encodedPosition.size(); ++i)
+		for (size_t i = 0; i < _encodedPosition.size(); ++i)
 		{
 			free(_encodedPosition[i]);
 			_encodedPosition[i]=nullptr;
@@ -90,19 +90,19 @@ public:
 		//	fprintf(stderr, "%s %d vs %d\n", "failure",_convertionTypeVector[0].size(),neighborValueArrayVector[0].size());
 
 		unsigned indexCenter=0;
-		for (int i =  _kernel->_dims.size()-1; i>=0 ; i--)
+		for (int i =  int(_kernel->_dims.size()-1); i>=0 ; i--)
 		{
 			indexCenter=indexCenter*_kernel->_dims[i]+_kernel->_dims[i]/2;
 		}
 
-		for (int i = 0; i < _convertionTypeVector.size(); ++i)
+		for (size_t i = 0; i < _convertionTypeVector.size(); ++i)
 		{
 
-			for (int j = 0; j < _convertionTypeVector[i].size(); ++j)
+			for (size_t j = 0; j < _convertionTypeVector[i].size(); ++j)
 			{
 				switch(_convertionTypeVector[i][j]){
 					case convertionType::P0:
-						for (int k = 0; k < neighborArrayVector.size(); ++k)
+						for (size_t k = 0; k < neighborArrayVector.size(); ++k)
 						{
 							unsigned indexInKernel=indexCenter;
 							if(_kernel->indexWithDelta(indexInKernel, indexCenter, neighborArrayVector[k]) && !std::isnan(neighborValueArrayVector[k][i]))
@@ -112,7 +112,7 @@ public:
 						}
 					break;
 					case convertionType::P1:
-						for (int k = 0; k < neighborArrayVector.size(); ++k)
+						for (size_t k = 0; k < neighborArrayVector.size(); ++k)
 						{
 							unsigned indexInKernel=indexCenter;
 							//fprintf(stderr, "%d ==> %f\n", _kernel->indexWithDelta(indexInKernel, indexCenter, neighborArrayVector[k]) && !std::isnan(neighborValueArrayVector[k][i]),neighborValueArrayVector[k][i]);
@@ -123,7 +123,7 @@ public:
 						}
 					break;
 					case convertionType::P2:
-						for (int k = 0; k < neighborArrayVector.size(); ++k)
+						for (size_t k = 0; k < neighborArrayVector.size(); ++k)
 						{
 							unsigned indexInKernel=indexCenter;
 							if(_kernel->indexWithDelta(indexInKernel, indexCenter, neighborArrayVector[k]) && !std::isnan(neighborValueArrayVector[k][i]))
@@ -142,16 +142,16 @@ public:
 		std::vector<float> delta;
 		//if(_completeTIs)
 		{
-			for (int p = 0; p < _convertionTypeVectorConstVector.size(); ++p)
+			for (size_t p = 0; p < _convertionTypeVectorConstVector.size(); ++p)
 			{
 				float sum=0;
-				for (int i = 0; i < _convertionTypeVectorConstVector[p].size(); ++i)
+				for (size_t i = 0; i < _convertionTypeVectorConstVector[p].size(); ++i)
 				{
-					for (int j = 0; j < _convertionTypeVectorConstVector[p][i].size(); ++j)
+					for (size_t j = 0; j < _convertionTypeVectorConstVector[p][i].size(); ++j)
 					{
 						switch(_convertionTypeVectorConstVector[p][i][j]){
 							case convertionType::P0:
-								for (int k = 0; k < neighborArrayVector.size(); ++k)
+								for (size_t k = 0; k < neighborArrayVector.size(); ++k)
 								{
 									unsigned indexInKernel=indexCenter;
 									if(_kernel->indexWithDelta(indexInKernel, indexCenter, neighborArrayVector[k]) && !std::isnan(neighborValueArrayVector[k][i]))
@@ -159,7 +159,7 @@ public:
 								}
 							break;
 							case convertionType::P1:
-								for (int k = 0; k < neighborArrayVector.size(); ++k)
+								for (size_t k = 0; k < neighborArrayVector.size(); ++k)
 								{
 									unsigned indexInKernel=indexCenter;
 									//fprintf(stderr, "%d ==> %f\n", _kernel->indexWithDelta(indexInKernel, indexCenter, neighborArrayVector[k]) && !std::isnan(neighborValueArrayVector[k][i]),neighborValueArrayVector[k][i]);
@@ -168,7 +168,7 @@ public:
 								}
 							break;
 							case convertionType::P2:
-								for (int k = 0; k < neighborArrayVector.size(); ++k)
+								for (size_t k = 0; k < neighborArrayVector.size(); ++k)
 								{
 									unsigned indexInKernel=indexCenter;
 									if(_kernel->indexWithDelta(indexInKernel, indexCenter, neighborArrayVector[k]) && !std::isnan(neighborValueArrayVector[k][i]))
@@ -204,7 +204,7 @@ public:
 		}
 
 		#pragma omp parallel for default(none) num_threads(_nbThreadOverTI) firstprivate(toUpdate,vectorSize,delta,moduleID) shared(updated, neighborArrayVector, convertedNeighborValueArrayVector, cummulatedVariablesCoeficient) 
-		for (int i = 0; i < vectorSize; ++i)
+		for (unsigned int i = 0; i < vectorSize; ++i)
 		{
 			if(toUpdate[i]){
 				updated[i]=_cdmV[moduleID][i]->candidateForPatern(neighborArrayVector, convertedNeighborValueArrayVector, cummulatedVariablesCoeficient,delta);
@@ -215,7 +215,7 @@ public:
 		std::fill(errors,errors+vectorSize*extendK,-INFINITY);
 
 		#pragma omp parallel for default(none) num_threads(_nbThreadOverTI) /*proc_bind(close)*/ firstprivate(seed, extendK,errors,encodedPosition,vectorSize,delta,moduleID,verbatimRecord,variableOfInterest) shared(updated, neighborArrayVector, convertedNeighborValueArrayVector, cummulatedVariablesCoeficient) 
-		for (int i = 0; i < vectorSize; ++i)
+		for (unsigned int i = 0; i < vectorSize; ++i)
 		{
 			float maxValue=delta.back();
 			if(updated[i])
@@ -228,7 +228,7 @@ public:
 				{
 					_cdmV[moduleID][i]->maskCroossErrorWithVariable(variableOfInterest);
 					#pragma omp simd
-					for (int j = 0; j < _cdmV[moduleID][i]->getErrorsArraySize(); ++j)
+					for (unsigned int j = 0; j < _cdmV[moduleID][i]->getErrorsArraySize(); ++j)
 					{
 						errosArray[j]=-std::fabs(errosArray[j]/(crossErrosArray[j]*crossErrosArray[j]*crossErrosArray[j]*crossErrosArray[j]));
 						if(crossErrosArray[j]==0.0f) errosArray[j]=-INFINITY;
@@ -267,7 +267,7 @@ public:
 				for (int j = 0; j <extendK ; ++j)
 				{
 					unsigned bestIndex=0;
-					for (int l = 1; l < _threadRatio*extendK; ++l)
+					for (unsigned int l = 1; l < _threadRatio*extendK; ++l)
 					{
 						if(localError[l] > localError[bestIndex]) bestIndex=l;
 					}
@@ -320,19 +320,19 @@ public:
 		//	fprintf(stderr, "%s %d vs %d\n", "failure",_convertionTypeVector[0].size(),neighborValueArrayVector[0].size());
 
 		unsigned indexCenter=0;
-		for (int i =  _kernel->_dims.size()-1; i>=0 ; i--)
+		for (int i =  int(_kernel->_dims.size()-1); i>=0 ; i--)
 		{
 			indexCenter=indexCenter*_kernel->_dims[i]+_kernel->_dims[i]/2;
 		}
 
-		for (int i = 0; i < _convertionTypeVector.size(); ++i)
+		for (size_t i = 0; i < _convertionTypeVector.size(); ++i)
 		{
 
-			for (int j = 0; j < _convertionTypeVector[i].size(); ++j)
+			for (size_t j = 0; j < _convertionTypeVector[i].size(); ++j)
 			{
 				switch(_convertionTypeVector[i][j]){
 					case convertionType::P0:
-						for (int k = 0; k < neighborArrayVector.size(); ++k)
+						for (size_t k = 0; k < neighborArrayVector.size(); ++k)
 						{
 							unsigned indexInKernel=indexCenter;
 							if(_kernel->indexWithDelta(indexInKernel, indexCenter, neighborArrayVector[k]) && !std::isnan(neighborValueArrayVector[k][i]))
@@ -342,7 +342,7 @@ public:
 						}						
 					break;
 					case convertionType::P1:
-						for (int k = 0; k < neighborArrayVector.size(); ++k)
+						for (size_t k = 0; k < neighborArrayVector.size(); ++k)
 						{
 							unsigned indexInKernel=indexCenter;
 							if(_kernel->indexWithDelta(indexInKernel, indexCenter, neighborArrayVector[k]) && !std::isnan(neighborValueArrayVector[k][i]))
@@ -352,7 +352,7 @@ public:
 						}						
 					break;
 					case convertionType::P2:
-						for (int k = 0; k < neighborArrayVector.size(); ++k)
+						for (size_t k = 0; k < neighborArrayVector.size(); ++k)
 						{
 							unsigned indexInKernel=indexCenter;
 							if(_kernel->indexWithDelta(indexInKernel, indexCenter, neighborArrayVector[k]) && !std::isnan(neighborValueArrayVector[k][i]))
@@ -371,16 +371,16 @@ public:
 		std::vector<float> delta;
 		//if(_completeTIs)
 		{
-			for (int p = 0; p < _convertionTypeVectorConstVector.size(); ++p)
+			for (size_t p = 0; p < _convertionTypeVectorConstVector.size(); ++p)
 			{
 				float sum=0;
-				for (int i = 0; i < _convertionTypeVectorConstVector[p].size(); ++i)
+				for (size_t i = 0; i < _convertionTypeVectorConstVector[p].size(); ++i)
 				{
-					for (int j = 0; j < _convertionTypeVectorConstVector[p][i].size(); ++j)
+					for (size_t j = 0; j < _convertionTypeVectorConstVector[p][i].size(); ++j)
 					{
 						switch(_convertionTypeVectorConstVector[p][i][j]){
 							case convertionType::P0:
-								for (int k = 0; k < neighborArrayVector.size(); ++k)
+								for (size_t k = 0; k < neighborArrayVector.size(); ++k)
 								{
 									unsigned indexInKernel=indexCenter;
 									if(_kernel->indexWithDelta(indexInKernel, indexCenter, neighborArrayVector[k]) && !std::isnan(neighborValueArrayVector[k][i]))
@@ -388,7 +388,7 @@ public:
 								}
 							break;
 							case convertionType::P1:
-								for (int k = 0; k < neighborArrayVector.size(); ++k)
+								for (size_t k = 0; k < neighborArrayVector.size(); ++k)
 								{
 									unsigned indexInKernel=indexCenter;
 									//fprintf(stderr, "%d ==> %f\n", _kernel->indexWithDelta(indexInKernel, indexCenter, neighborArrayVector[k]) && !std::isnan(neighborValueArrayVector[k][i]),neighborValueArrayVector[k][i]);
@@ -397,7 +397,7 @@ public:
 								}
 							break;
 							case convertionType::P2:
-								for (int k = 0; k < neighborArrayVector.size(); ++k)
+								for (size_t k = 0; k < neighborArrayVector.size(); ++k)
 								{
 									unsigned indexInKernel=indexCenter;
 									if(_kernel->indexWithDelta(indexInKernel, indexCenter, neighborArrayVector[k]) && !std::isnan(neighborValueArrayVector[k][i]))
@@ -416,7 +416,7 @@ public:
 		}
 
 		#pragma omp parallel for default(none) num_threads(_nbThreadOverTI) firstprivate(vectorSize,delta,moduleID) shared(updated, neighborArrayVector, convertedNeighborValueArrayVector, cummulatedVariablesCoeficient) 
-		for (int i = 0; i < vectorSize; ++i)
+		for (unsigned int i = 0; i < vectorSize; ++i)
 		{
 			updated[i]=_cdmV[moduleID][i]->candidateForPatern(neighborArrayVector, convertedNeighborValueArrayVector, cummulatedVariablesCoeficient,delta);
 		}
@@ -425,7 +425,7 @@ public:
 		std::fill(errors,errors+vectorSize*extendK,-INFINITY);
 
 		#pragma omp parallel for default(none) num_threads(_nbThreadOverTI) firstprivate(extendK,errors,encodedPosition,vectorSize,delta,moduleID) shared(updated, neighborArrayVector, convertedNeighborValueArrayVector, cummulatedVariablesCoeficient) 
-		for (int i = 0; i < vectorSize; ++i)
+		for (unsigned int i = 0; i < vectorSize; ++i)
 		{
 			if(updated[i])
 			{
@@ -436,7 +436,7 @@ public:
 				{
 					_cdmV[moduleID][i]->maskCroossError();
 					#pragma omp simd
-					for (int j = 0; j < _cdmV[moduleID][i]->getErrorsArraySize(); ++j)
+					for (unsigned int j = 0; j < _cdmV[moduleID][i]->getErrorsArraySize(); ++j)
 					{
 						errosArray[j]=-std::fabs(errosArray[j]/(crossErrosArray[j]*crossErrosArray[j]*crossErrosArray[j]*crossErrosArray[j]));
 						if(crossErrosArray[j]==0.0f) errosArray[j]=-INFINITY;
